fixes 1201143 - consolidating use of STOPWORDS list
