+++ airmozilla/base/utils.py
@@ -23,6 +23,20 @@ from django.contrib.staticfiles.storage import staticfiles_storage
 from airmozilla.base.akamai_token_v2 import AkamaiToken
 
 
+STOPWORDS = (
+    "a able about across after all almost also am among an and "
+    "any are as at be because been but by can cannot could dear "
+    "did do does either else ever every for from get got had has "
+    "have he her hers him his how however i if in into is it its "
+    "just least let like likely may me might most must my "
+    "neither no nor not of off often on only or other our own "
+    "rather said say says she should since so some than that the "
+    "their them then there these they this tis to too twas us "
+    "wants was we were what when where which while who whom why "
+    "will with would yet you your".split()
+)
+
+
 def roughly(number, variance_percentage=20):
     """return a number that is roughly what you inputted but
     slightly smaller or slightly bigger.
